+++ Sources/KeyboardKit/Proxy/UITextDocumentProxy+FullDocumentContext.swift
@@ -15,6 +15,9 @@ public extension UITextDocumentProxy {
     /**
      Whether or not this proxy is currently reading the full
      document context.
+
+     The full document context reading capabilities are only
+     available in KeyboardKit Pro.
      */
     var isReadingFullDocumentContext: Bool {
         get { FullDocumentContextReaderState.isReadingFullDocumentContext }